requirements.txt
grpcio
grpcio-tools
pyyaml
redis
asyncio
futures
docker